Move training application form related classes to a separate namespace
